refactor: No need for stylus
