+++ packages/cozy-scanner/src/CategoryGridItem.jsx
@@ -4,7 +4,8 @@ import classNames from 'classnames'
 import palette from 'cozy-ui/transpiled/react/palette'
 import { Icon, IconStack } from 'cozy-ui/transpiled/react'
 
-import styles from './styles.styl'
+import styles from './stylesheet.css'
+console.log('styles', styles)
 
 const CategoryGridItem = ({ isSelected, icon, theme, label }) => {
   return (
