+++ packages/cozy-scanner/src/DocumentQualification.jsx
@@ -14,7 +14,7 @@ import { themes } from './DocumentTypeData'
 import { getItemById } from './DocumentTypeDataHelpers'
 import { getItemsByCategory } from './DocumentTypeDataHelpers'
 import GridItem from 'cozy-ui/transpiled/react/Labs/GridItem'
-import styles from './styles.styl'
+import styles from './stylesheet.css'
 
 const fileExtension = '.jpg'
 const idFileInput = 'filename_input'
